写入文件告一段落todo

+++ src/renderer/pages/db/list.vue
@@ -2,9 +2,10 @@
     <div>
         <h1>数据库列表</h1>
         <hr>
-        <el-table :data="databases" style="width: 100%" height="400" class="table table-striped table-bordered">
+        <el-table :data="databases" style="width: 100%" height="400" class="table table-striped table-bordered"
+                  v-loading="loading">
             <el-table-column prop="id" label="id" width="40"/>
-            <el-table-column prop="name" label="简称" width="80"/>
+            <el-table-column prop="name" label="简称" width="110"/>
             <el-table-column prop="type" label="类型" width="60"/>
             <el-table-column prop="host" label="地址"/>
             <el-table-column prop="database" label="数据库" width="100"/>
@@ -27,7 +28,8 @@
     name: 'list',
     data () {
       return {
-        databases: []
+        databases: [],
+        loading: true
       }
     },
     mounted () {
@@ -39,7 +41,9 @@
        * @return {Promise<void>}
        */
       async getDatabases () {
+        this.loading = true
         this.databases = await Database.findAll()
+        this.loading = false
       },
       /**
        * 编辑数据库
@@ -60,6 +64,11 @@
        * @param {Database} database
        */
       async remove (database) {
+        let confirm = await this.$confirm(`确认删除【${database.name}】吗？`, '提示', {}).catch(e => false)
+        if (confirm !== 'confirm') {
+          return false
+        }
+
         await database.destroy()
         this.getDatabases()
       }
